cleanup and add extra validation
