+++ x/cronos/types/params.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"fmt"
+	"strings"
 
 	yaml "gopkg.in/yaml.v2"
 
@@ -12,6 +13,8 @@ var (
 	// KeyIbcCroDenom is store's key for the IBC Cro denomination
 	KeyIbcCroDenom = []byte("IbcCroDenom")
 )
+
+const IbcCroDenomDefaultValue = "ibc/6B5A664BF0AF4F71B2F0BAA33141E2F1321242FBD5D19762F541EC971ACB0865"
 
 // ParamKeyTable returns the parameter key table.
 func ParamKeyTable() paramtypes.KeyTable {
@@ -28,13 +31,13 @@ func NewParams(ibcCroDenom string) Params {
 // DefaultParams is the default parameter configuration for the cronos module
 func DefaultParams() Params {
 	return Params{
-		IbcCroDenom: "ibc/6B5A664BF0AF4F71B2F0BAA33141E2F1321242FBD5D19762F541EC971ACB0865",
+		IbcCroDenom: IbcCroDenomDefaultValue,
 	}
 }
 
 // Validate all cronos module parameters
 func (p Params) Validate() error {
-	return validateIsString(p.IbcCroDenom)
+	return validateIsIbcDenom(p.IbcCroDenom)
 }
 
 // String implements the fmt.Stringer interface
@@ -46,14 +49,18 @@ func (p Params) String() string {
 // ParamSetPairs implements params.ParamSet
 func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
 	return paramtypes.ParamSetPairs{
-		paramtypes.NewParamSetPair(KeyIbcCroDenom, &p.IbcCroDenom, validateIsString),
+		paramtypes.NewParamSetPair(KeyIbcCroDenom, &p.IbcCroDenom, validateIsIbcDenom),
 	}
 }
 
-func validateIsString(i interface{}) error {
-	_, ok := i.(string)
+func validateIsIbcDenom(i interface{}) error {
+	s, ok := i.(string)
 	if !ok {
 		return fmt.Errorf("invalid parameter type: %T", i)
+	}
+
+	if s != "" && (!strings.HasPrefix(s, "ibc/") || len(s) != 68) {
+		return fmt.Errorf("invalid ibc denom: %T", i)
 	}
 	return nil
 }
